+++ app/src/main/java/com/mobven/fitai/home/screens/SportOftenFragment.kt
@@ -1,6 +1,6 @@
-package com.mobven.fitai.login.sign_up.screens
+package com.mobven.fitai.home.screens
 
-
+import androidx.viewpager2.widget.ViewPager2
 import com.mobven.fitai.R
 import com.mobven.fitai.base.BaseFragment
 import com.mobven.fitai.databinding.FragmentGoalsBinding
@@ -8,42 +8,45 @@ import com.mobven.fitai.login.sign_up.adapter.SignUpListAdapter
 import com.mobven.fitai.login.sign_up.model.SignUpSelectorItem
 import com.mobven.fitai.util.enums.SignUpSelectorType
 
-class GoalsFragment : BaseFragment<FragmentGoalsBinding>(FragmentGoalsBinding::inflate) {
+
+class SportOftenFragment : BaseFragment<FragmentGoalsBinding>(FragmentGoalsBinding::inflate) {
 
     private val adapter = SignUpListAdapter()
-    override fun observeUi() {
 
-        val goalsList = listOf(
+    override fun observeUi() {
+        val sportOftenList = listOf(
             SignUpSelectorItem(
-                title = getString(R.string.calorie_control),
+                title = getString(R.string.as_much_as_offered),
                 isSelected = false,
                 type = SignUpSelectorType.RADIO
             ),
             SignUpSelectorItem(
-                title = getString(R.string.muscle_gain),
+                title = getString(R.string.on_two_times_week),
                 isSelected = false,
                 type = SignUpSelectorType.RADIO
             ),
             SignUpSelectorItem(
-                title = getString(R.string.weight_gain),
+                title = getString(R.string.three_four_times_week),
                 isSelected = false,
                 type = SignUpSelectorType.RADIO
             ),
             SignUpSelectorItem(
-                title = getString(R.string.weight_loss),
+                title = getString(R.string.five_six_times_week),
                 isSelected = false,
                 type = SignUpSelectorType.RADIO
             )
-        )
+        ).reversed()
 
-        adapter.submitList(goalsList)
+        adapter.submitList(sportOftenList)
         binding.rvGoals.adapter = adapter
 
         binding.btnGoalsContinue.setOnClickListener {
-//            val currentItem = requireActivity().findViewById<ViewPager2>(R.id.vp_nutrition).currentItem
-//            val nextItem = currentItem + 1
-//            requireActivity().findViewById<ViewPager2>(R.id.vp_nutrition).setCurrentItem(nextItem, true)
+            val currentItem =
+                requireActivity().findViewById<ViewPager2>(R.id.vp_training).currentItem
+            val nextItem = currentItem + 1
+            requireActivity().findViewById<ViewPager2>(R.id.vp_training)
+                .setCurrentItem(nextItem, true)
         }
-
     }
+
 }
