+++ app/src/main/java/com/mobven/fitai/home/training/TrainingPagerAdapter.kt
@@ -3,12 +3,12 @@ package com.mobven.fitai.home.training
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.mobven.fitai.login.sign_up.screens.GoalsFragment
-import com.mobven.fitai.login.sign_up.screens.HealthProblemFragment
-import com.mobven.fitai.login.sign_up.screens.PlanCreatingFragment
-import com.mobven.fitai.login.sign_up.screens.PreferredSportFragment
-import com.mobven.fitai.login.sign_up.screens.SportBodyPartFragment
-import com.mobven.fitai.login.sign_up.screens.SportOftenFragment
+import com.mobven.fitai.home.screens.GoalsFragment
+import com.mobven.fitai.home.screens.HealthProblemFragment
+import com.mobven.fitai.home.screens.PlanCreatingFragment
+import com.mobven.fitai.home.screens.PreferredSportFragment
+import com.mobven.fitai.home.screens.SportBodyPartFragment
+import com.mobven.fitai.home.screens.SportOftenFragment
 
 class TrainingPagerAdapter(
     fragmentActivity: FragmentActivity
