+++ app/src/main/res/layout/fragment_diet_type.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.screens.DietTypeFragment">
+    tools:context=".home.screens.DietTypeFragment">
 
     <TextView
         android:id="@+id/diet_type_title"
