+++ app/src/main/res/layout/fragment_food_allergies.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.screens.FoodAllergiesFragment">
+    tools:context=".home.screens.FoodAllergiesFragment">
 
     <TextView
         android:id="@+id/food_allergies_title"
