+++ app/src/main/res/layout/fragment_health_problem.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.screens.HealthProblemFragment">
+    tools:context=".home.screens.HealthProblemFragment">
 
     <TextView
         android:id="@+id/health_problem_title"
