+++ app/src/main/res/layout/fragment_other_health_problem.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.screens.OtherHealthProblemFragment">
+    tools:context=".home.screens.OtherHealthProblemFragment">
 
     <TextView
         android:id="@+id/other_health_problem_title"
