+++ app/src/main/res/layout/fragment_preferred_sport.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.sign_up.screens.PreferredSportFragment">
+    tools:context=".home.screens.PreferredSportFragment">
 
     <TextView
         android:id="@+id/preferred_sport_title"
