+++ app/src/main/res/layout/fragment_sport_body_part.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:background="@color/white"
-    tools:context=".login.sign_up.screens.SportBodyPartFragment">
+    tools:context=".home.screens.SportBodyPartFragment">
 
     <TextView
         android:id="@+id/body_part_title"
